feat: 按Arduino库目录结构要求适配工程目录
