+++ examples/01_device/main.c
@@ -4,9 +4,9 @@
 #define NODA_HEARTBEAT_MILLIS 1000
 
 #include <noda.h>
-#include <noda_log.h>
+#include <noda/log.h>
 
-#include "noda_io.h"
+#include "noda/io.h"
 
 /*************************************************************************
   * 系统默认使用总线(BUS)管理协议复用，若项目不需要启用，请使用此声明
