+++ noda-library/library.properties
@@ -1,0 +1,10 @@
+name=noda
+version=0.0.1
+author=[name/nickname of the authors and their email addresses (not mandatory) separated by comma (,)]
+maintainer=Tiwater
+sentence=[a sentence explaining the purpose of the library]
+paragraph=[a longer description of the library. The value of sentence will be prepended to this, so you should start by writing the second sentence here]
+category=Other
+url=https://github.com/tiwater/noda
+architectures=*
+includes=noda.h
