+++ noda-library/src/core/device_center.c
@@ -1,7 +1,7 @@
-#include "noda_device_center.h"
-#include "noda_task.h"
-#include "noda_time.h"
-#include "noda_log.h"
+#include "noda/device_center.h"
+#include "noda/task.h"
+#include "noda/time.h"
+#include "noda/log.h"
 
 static noda_task_t* s_task;
