+++ noda-library/src/core/noda.c
@@ -1,8 +1,8 @@
-#include "noda_internal.h"
-#include "noda_time.h"
-#include "noda_bus_center.h"
-#include "noda_device_center.h"
-#include "noda_log.h"
+#include "noda/internal.h"
+#include "noda/time.h"
+#include "noda/bus_center.h"
+#include "noda/device_center.h"
+#include "noda/log.h"
 
 static int g_ladder;
