+++ noda-library/src/noda/bus.h
@@ -1,7 +1,7 @@
 #ifndef __NODA_BUS_H
 #define __NODA_BUS_H
 
-#include "noda_common.h"
+#include "noda/common.h"
 
 #ifdef __cplusplus
 extern "C" {
