+++ noda-library/src/noda/common.h
