+++ noda-library/src/noda/device.h
@@ -1,7 +1,7 @@
 #ifndef __NODA_DEVICE_H
 #define __NODA_DEVICE_H
 
-#include "noda_common.h"
+#include "noda/common.h"
 
 #ifdef __cplusplus
 extern "C" {
