+++ noda-library/src/noda/task.h
@@ -1,7 +1,7 @@
 #ifndef __NODA_TASK_H
 #define __NODA_TASK_H
 
-#include "noda_common.h"
+#include "noda/common.h"
 
 #ifdef __cplusplus
 extern "C" {
