+++ platforms/esp32/examples/02_bus/main/silan_sc7a20.c
@@ -4,7 +4,7 @@
   ************************************************************************/
 
 #include "silan_sc7a20.h"
-#include "noda_bus_center.h"
+#include "noda/bus_center.h"
 
 /**********sc7a20**********/
 #define IIC_ADDR        0x19
