+++ platforms/esp32/src/io.c
@@ -3,11 +3,10 @@
   * @generate date: 2022-10-06 21:35:04
   ************************************************************************/
 
-#include "noda_io.h"
+#include "noda/io.h"
+#include "noda/log.h"
 
 #include <driver/gpio.h>
-
-#include "noda_log.h"
 
 /*************************************************************************
   * noda_io_open：noda_io必须实现的类成员函数，负责设备“打开”操作，
