+++ platforms/unix/src/time.c
@@ -1,4 +1,4 @@
-#include "noda_time.h"
+#include "noda/time.h"
 #include <sys/select.h>
 
 void noda_delay(clock_t ms) {
